packages/pa-core/pyproject.toml
[project]
name = "pa-core"
version = "0.1.0"
description = "PA shared utilities — config, CLI runner, logging"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0", "python-dotenv>=1.0"]

[project.scripts]
pa-core = "pa_core.setup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
